backend/requirements.txt
fastapi==0.110.0
uvicorn==0.23.2
pydantic>=2.7.4,<3.0.0
python-dotenv==1.0.0
mysql-connector-python==8.1.0
python-multipart==0.0.6
numpy==1.26.4
google-generativeai>=0.8.0,<0.9.0
google-ai-generativelanguage>=0.4.0
protobuf>=4.21.12
langchain>=0.3.0
langchain-core>=0.3.51,<1.0.0
langchain-google-genai>=2.0.0
langsmith>=0.0.83

# Semantic search dependencies
sentence-transformers>=2.2.2
scikit-learn>=1.2.2